Рассписание готово

+++ Shedule_04/ModalsForm/SheduleGroupForm.cs
@@ -44,7 +44,154 @@ namespace Shedule_04.ModalsForm
         private void SheduleGroupForm_Load(object sender, EventArgs e)
         {
             comboLoad();
-            tableLoad();
+            //tableLoad();
+            tableLoad2();
+        }
+
+        private void tableLoad2()
+        {
+            #region ADD Rows
+            //add empty rows
+            dataGridView1.Rows.Clear();
+            for (int i = 0; i < 43; i++)
+            {
+                dataGridView1.Rows.Add();
+            }
+
+            // add faculty, group, days, times
+            dataGridView1[3, 0].Value = "" + semester + " семестр " + year + " года \n" +
+                "Группа " + group + "";
+            int k = 0;
+            for (int i = 1; i < 43; i += 6)
+            {
+                dataGridView1[1, i].Value = days[k];
+                k++;
+            }
+            for (int i = 1; i < 43; i++)
+            {
+                if (k < 6)
+                {
+                    dataGridView1[2, i].Value = times[k];
+                    k++;
+                }
+                else
+                {
+                    k = 0;
+                    dataGridView1[2, i].Value = times[k];
+                    k++;
+                }
+            }
+            int s = 0, t = 0;
+            for (int i = 1; i < 43; i++)
+            {
+                //dataGridView1[3, i].Value = s.ToString();
+                if (t > 5)
+                {
+                    t = 0;
+                    s++;
+                    dataGridView1[6, i].Value = s.ToString();
+                    t++;
+                }
+                else
+                {
+                    dataGridView1[6, i].Value = s.ToString();
+                    t++;
+                }
+            }
+            #endregion
+            #region ADD Items
+            // Получаем все ИД квантов расписания
+            List<string> ids = new List<string>();
+            try
+            {
+                string gr = @"select id_group From groups where group_name = '" + group + "'";
+                SqlCommand table1 = new SqlCommand(gr, connect);
+                connect.Open();
+                SqlDataReader reader1 = table1.ExecuteReader();
+                reader1.Read();
+                string idGroup = reader1[0].ToString();
+                connect.Close();
+
+                string getIds = @"SELECT shed_time, id_group
+                                FROM shedule_table JOIN groups on fk_group = id_group
+                                WHERE year = '" + year + "' AND semester = '" + semester + "' AND id_group = '" + idGroup + "'";
+                SqlCommand table = new SqlCommand(getIds, connect);
+                connect.Open();
+                SqlDataReader reader = table.ExecuteReader();
+                while (reader.Read())
+                {
+                    ids.Add(reader[0].ToString());
+                }
+                connect.Close();
+
+            }
+            catch (SqlException ex)
+            {
+                connect.Close();
+                MessageBox.Show(ex.Number.ToString(), "Неизвестная ошибка.");
+            }
+
+            //Заполняем таблицу с расписанием
+            try
+            {
+                string[] queryIds = ids.ToArray();
+                string id = "(";
+
+                for (int j = 0; j < queryIds.Length; j++)
+                {
+                    if (j != queryIds.Length - 1)
+                    {
+                        id += "'" + ids[j] + "',";
+                    }
+                    else
+                    {
+                        id += "'" + ids[j] + "')";
+                    }
+                }
+
+
+                string querieAll = @"select id_shTime, day, task_time, subject_name, surname, classroom_name 
+	                               from shedule_time JOIN subject on fk_subject = id_subject JOIN classroom on fk_classroom = id_classroom JOIN lecturer on fk_lecturer = id_lecturer
+	                               where id_shTime in " + id + " order by day, task_time ";
+
+                SqlCommand table = new SqlCommand(querieAll, connect);
+
+                connect.Open();
+
+                SqlDataReader reader = table.ExecuteReader();
+
+                while (reader.Read())
+                {
+                    // Find row
+
+                    for (int jj = 1; jj < 43; jj += 6)
+                    {
+                        if (reader[1].ToString() == dataGridView1[6, jj].Value.ToString())
+                        {
+                            for (int kk = jj; kk < (jj += 6); kk++)
+                            {
+
+                                if (reader[2].ToString() == dataGridView1[2, kk].Value.ToString())
+                                {
+                                    dataGridView1[0, kk].Value = reader[0];
+                                    dataGridView1[3, kk].Value = reader[3];
+                                    dataGridView1[4, kk].Value = reader[4];
+                                    dataGridView1[5, kk].Value = reader[5];
+                                    break;
+                                }
+                            }
+                        }
+                    }
+                }
+                reader.Close();
+                connect.Close();
+            }
+            catch (SqlException ex)
+            {
+                connect.Close();
+                MessageBox.Show(ex.Number.ToString(), "Неизвестная ошибка.");
+            }
+            #endregion
         }
 
         private void btn_add_Click(object sender, EventArgs e)
@@ -166,12 +313,12 @@ namespace Shedule_04.ModalsForm
                     }
                 }
 
-                
+
                 // Если все ок, то разрешаем добавить
                 if (idTask == null && idLect == null && idClass == null)
                 {
-                    
-                   
+
+
                     string addTime = @"Insert Into shedule_time (day, task_time, fk_subject, fk_classroom, fk_lecturer) VALUES
                                     ('" + combo_day.SelectedIndex + "', " +
                                     "'" + combo_time.SelectedItem + "', " +
@@ -192,7 +339,8 @@ namespace Shedule_04.ModalsForm
                     insert2.ExecuteNonQuery();
                 }
                 connect.Close();
-                tableLoad();
+                //tableLoad();
+                tableLoad2();
             }
             catch (SqlException ex)
             {
@@ -203,7 +351,7 @@ namespace Shedule_04.ModalsForm
 
         private void comboLoad()
         {
-                       
+
             combo_day.DataSource = days;
             combo_time.DataSource = times;
 
@@ -254,7 +402,7 @@ namespace Shedule_04.ModalsForm
 
         private void tableLoad()
         {
-            
+
             // Получаем все ИД квантов расписания
             List<string> ids = new List<string>();
             try
@@ -318,17 +466,17 @@ namespace Shedule_04.ModalsForm
                     int i = 0;
                     dataGridView1.Rows.Clear();
 
-                    
+
                     while (reader.Read())
                     {
 
                         dataGridView1.Rows.Add();
                         dataGridView1[0, i].Value = reader[0]; //id_row
-                        
+
                         // Сортировка дня недели и отображение
-                        int nameOfDay = Int32.Parse( reader[1].ToString());
+                        int nameOfDay = Int32.Parse(reader[1].ToString());
                         dataGridView1[1, i].Value = days[nameOfDay];
-                       
+
                         dataGridView1[2, i].Value = reader[2];
                         dataGridView1[3, i].Value = reader[3];
                         dataGridView1[4, i].Value = reader[4];
@@ -338,7 +486,7 @@ namespace Shedule_04.ModalsForm
                     reader.Close();
                     connect.Close();
 
-                    int j = dataGridView1.Rows.Count-1;
+                    int j = dataGridView1.Rows.Count - 1;
                     while (j != 0)
                     {
                         if (j != 0 && dataGridView1[1, j].Value.ToString() == dataGridView1[1, j - 1].Value.ToString())
@@ -360,38 +508,43 @@ namespace Shedule_04.ModalsForm
                 MessageBox.Show(ex.Number.ToString(), "Неизвестная ошибка.");
             }
         }
-       
+
         private void reloadLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
         {
-            tableLoad();
+            //tableLoad();
+            tableLoad2();
         }
 
         private void deleteLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
         {
             int row = dataGridView1.SelectedRows[0].Index;
-            string ids = dataGridView1[0, row].Value.ToString();
-
-            string delete = @"DELETE FROM shedule_time 
+            if (dataGridView1[0, row].Value != null)
+            {
+                string ids = dataGridView1[0, row].Value.ToString();
+                string delete = @"DELETE FROM shedule_time 
                             WHERE id_shTime = '" + ids + "'";
 
-            if (MessageBox.Show("Вы действительно хотите удалить выбранные записи? Данная операция необратима.", "Удаление", MessageBoxButtons.YesNo) == DialogResult.Yes)
-            {
-                try
+                if (MessageBox.Show("Вы действительно хотите удалить выбранные записи? Данная операция необратима.", "Удаление", MessageBoxButtons.YesNo) == DialogResult.Yes)
                 {
-                    SqlCommand table = new SqlCommand(delete, connect);
+                    try
+                    {
+                        SqlCommand table = new SqlCommand(delete, connect);
 
-                    connect.Open();
-                    SqlDataReader reader2 = table.ExecuteReader();
-                    reader2.Close();
-                    connect.Close();
-                    tableLoad();
-                }
-                catch (SqlException ex)
-                {
-                    connect.Close();
-                    MessageBox.Show(ex.Number.ToString(), "Неизвестная ошибка.");
+                        connect.Open();
+                        SqlDataReader reader2 = table.ExecuteReader();
+                        reader2.Close();
+                        connect.Close();
+                        //tableLoad();
+                        tableLoad2();
+                    }
+                    catch (SqlException ex)
+                    {
+                        connect.Close();
+                        MessageBox.Show(ex.Number.ToString(), "Неизвестная ошибка.");
+                    }
                 }
             }
         }
     }
+
 }
